Rename flow_boundaries to boundaries
