+++ fehmtk/preprocessors/__init__.py
@@ -1,3 +1,3 @@
-from .flow_boundaries import generate_flow_boundaries, generate_heat_flux_boundaries
+from .boundaries import generate_flow_boundaries, generate_heat_flux_boundaries
 from .hydrostatic_pressure import generate_hydrostatic_pressure
 from .rock_properties import generate_rock_properties
